[FEAT] 유저셀 수정 (#67)
