+++ FoodBowl/Screens/Feed/Cell/StoreSearchTableViewCell.swift
@@ -1,8 +1,8 @@
 //
-//  StoreResultTableViewCell.swift
+//  StoreSearchTableViewCell.swift
 //  FoodBowl
 //
-//  Created by COBY_PRO on 2023/01/18.
+//  Created by Coby Kim on 2023/01/13.
 //
 
 import UIKit
@@ -10,7 +10,7 @@ import UIKit
 import SnapKit
 import Then
 
-final class StoreResultTableViewCell: BaseTableViewCell {
+final class StoreSearchTableViewCell: BaseTableViewCell {
     // MARK: - property
 
     let storeNameLabel = UILabel().then {
@@ -18,7 +18,7 @@ final class StoreResultTableViewCell: BaseTableViewCell {
         $0.textColor = .mainText
     }
 
-    let storeFeedLabel = UILabel().then {
+    let storeAdressLabel = UILabel().then {
         $0.font = UIFont.preferredFont(forTextStyle: .caption1, weight: .light)
         $0.textColor = .subText
     }
@@ -31,14 +31,14 @@ final class StoreResultTableViewCell: BaseTableViewCell {
     // MARK: - func
 
     override func render() {
-        contentView.addSubviews(storeNameLabel, storeFeedLabel, storeDistanceLabel)
+        contentView.addSubviews(storeNameLabel, storeAdressLabel, storeDistanceLabel)
 
         storeNameLabel.snp.makeConstraints {
             $0.leading.equalToSuperview().inset(20)
             $0.top.equalToSuperview().inset(12)
         }
 
-        storeFeedLabel.snp.makeConstraints {
+        storeAdressLabel.snp.makeConstraints {
             $0.leading.equalToSuperview().inset(20)
             $0.bottom.equalToSuperview().inset(12)
         }
