--- FoodBowl/Screens/Search/Cell/UserResultTableViewCell.swift
@@ -1,68 +1,0 @@
-//
-//  UserResultTableViewCell.swift
-//  FoodBowl
-//
-//  Created by COBY_PRO on 2023/01/18.
-//
-
-import UIKit
-
-import SnapKit
-import Then
-
-final class UserResultTableViewCell: BaseTableViewCell {
-    // MARK: - property
-
-    lazy var userImageView = UIImageView().then {
-        $0.backgroundColor = .gray
-        $0.layer.cornerRadius = 20
-        $0.layer.masksToBounds = true
-    }
-
-    let userNameLabel = UILabel().then {
-        $0.font = UIFont.preferredFont(forTextStyle: .subheadline, weight: .medium)
-        $0.textColor = .mainText
-        $0.text = "홍길동"
-    }
-
-    let userFollowerLabel = UILabel().then {
-        $0.font = UIFont.preferredFont(forTextStyle: .footnote, weight: .light)
-        $0.textColor = .subText
-        $0.text = "팔로워 100명"
-    }
-
-    let followButton = FollowButton()
-
-    // MARK: - func
-
-    override func render() {
-        contentView.addSubviews(userImageView, userNameLabel, userFollowerLabel, followButton)
-
-        userImageView.snp.makeConstraints {
-            $0.leading.equalToSuperview().inset(20)
-            $0.centerY.equalToSuperview()
-            $0.width.height.equalTo(40)
-        }
-
-        userNameLabel.snp.makeConstraints {
-            $0.leading.equalTo(userImageView.snp.trailing).offset(16)
-            $0.top.equalToSuperview().inset(12)
-        }
-
-        userFollowerLabel.snp.makeConstraints {
-            $0.leading.equalTo(userImageView.snp.trailing).offset(16)
-            $0.top.equalTo(userNameLabel.snp.bottom).offset(4)
-        }
-
-        followButton.snp.makeConstraints {
-            $0.trailing.equalToSuperview().inset(20)
-            $0.centerY.equalToSuperview()
-            $0.width.equalTo(60)
-            $0.height.equalTo(30)
-        }
-    }
-    
-    override func configUI() {
-        backgroundColor = .clear
-    }
-}
