+++ FoodBowl/Screens/Search/SearchUserResultViewController.swift
@@ -14,7 +14,7 @@ final class SearchUserResultViewController: BaseViewController {
     // MARK: - property
 
     private lazy var userResultTableView = UITableView().then {
-        $0.register(UserResultTableViewCell.self, forCellReuseIdentifier: UserResultTableViewCell.className)
+        $0.register(UserInfoTableViewCell.self, forCellReuseIdentifier: UserInfoTableViewCell.className)
         $0.delegate = self
         $0.dataSource = self
         $0.separatorInset = UIEdgeInsets(top: 0, left: 0, bottom: 20, right: 0)
@@ -42,7 +42,7 @@ extension SearchUserResultViewController: UITableViewDataSource, UITableViewDele
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: UserResultTableViewCell.className, for: indexPath) as? UserResultTableViewCell else { return UITableViewCell() }
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: UserInfoTableViewCell.className, for: indexPath) as? UserInfoTableViewCell else { return UITableViewCell() }
 
         cell.selectionStyle = .none
